Move validation test to validationTests.py
Put all plugins to be run into a separate file, to make it easier
to add additional tests. Add in some documentation in the process
and document a way in which a plugin can take additional parameters.
At this point the additional parameters are only a potential but the
documentation explains how to make use of them.